+++ packages/at_client/lib/src/crypto/crypto.dart
@@ -20,6 +20,7 @@ export 'package:at_client/src/crypto/nskey/nskey_private_filing.dart';
 export 'package:at_client/src/crypto/nskey/nskey_key_ring.dart';
 export 'package:at_client/src/crypto/nskey/nskey_provider.dart';
 export 'package:at_client/src/crypto/nskey/nskey_resolver.dart';
+export 'package:at_client/src/crypto/nskey/nskey_rotation.dart';
 export 'package:at_client/src/crypto/nskey/pq_signing_chain.dart';
 export 'package:at_client/src/crypto/nskey/pq_signing_root.dart';
 export 'package:at_client/src/crypto/nskey/published_nskey_key_ring.dart';